+++ src/Routes.js
@@ -5,8 +5,8 @@ import AuthenticatedRoute from './components/AuthenticatedRoute';
 import {USER_ROLES} from './libs/utils';
 import ManageDac from './pages/manage_dac/ManageDac';
 import AdminManageUsers from './pages/AdminManageUsers';
-import DataAccessRequestApplication_old from './pages/DataAccessRequestApplication_old';
-import DataAccessRequestApplication from './pages/DataAccessRequestApplication';
+import DataAccessRequestApplication_old from './pages/dar_application/DataAccessRequestApplication_old';
+import DataAccessRequestApplication from './pages/dar_application/DataAccessRequestApplication';
 import DatasetCatalog from './pages/DatasetCatalog';
 import DACDatasets from './pages/DACDatasets';
 import DatasetRegistration from './pages/DatasetRegistration';
